COM-2471 - Refacto types
