+++ src/types/DateTimeType.ts
@@ -1,0 +1,1 @@
+export type ModeType = 'date' | 'time';
